3d_model_diff/requirements.txt
matplotlib==3.5.3
numpy==1.21.6
open3d==0.17.0
pandas==1.3.5
Pillow==10.0.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
